+++ src/polyester/__init__.py
@@ -12,7 +12,7 @@ from polyester.errors import (
     PolyesterValidationError,
 )
 
-__version__ = "0.1.0a4"
+__version__ = "0.1.0a5"
 
 __all__ = [
     "AsyncPolyester",